moved and renamed components into own folders
