+++ frontend/src/components/AppCarousel/AppCarousel.css
