+++ frontend/src/pages/favorite-apps.js
@@ -1,4 +1,5 @@
 import React from "react"
+import "bootstrap/dist/css/bootstrap.min.css"
 
 import Layout from "../components/Layout/Layout"
 import SEO from "../components/seo"
